Rename depth slider component file to kebap case
