+++ frontend/src/components/main-page/graph-page/depth-slider.tsx
